Homework5 (#3)
* Add Task1

* add homework2

* add task

* add task

* выливка домашки

* добавлены задачи

* еще две задачи

* и еще две задачки, достала уже наверняка

* add hw5-1

* homework5

+++ src/main/java/homeworks_dz5/hw_task10.java
@@ -18,7 +18,7 @@ public class hw_task10 {
         if (operation.equals("*")) {
             System.out.println("x * y = " + (x * y));}
         if (operation.equals("/") && y != 0) {
-            System.out.println("x / y = " + (x / y));}
+            System.out.println("x / y = " + ((float)x / (float)y));}
         else System.out.println("Enter valid data");
-    }
+        }
 }
